move inference into a specific classifier package
